requirements.txt
python-telegram-bot==20.0
deep-translator==1.11.2
python-dotenv==1.0.0
flask==2.3.0